Add text_report family (G5) + migrate washington (per-candidate PDF report)
